Refactor the API and address comments
